requirements.txt
flask==3.0.3
flask-cors==4.0.1
pywebview==5.1.0
python-Levenshtein==0.25.1
pyinstaller==6.12.0
watchdog==3.0.0
